Update build action authorization and recompile button, set gRPC message size limits
This commit updates the GameProjectsController to add authorization
checks before building a game project and ensures the correct status is\
rendered. It also enhances the recompile button in the HTML game
compiles view to use turbo frames. Additionally, it sets explicit gRPC\
message size limits in the crew_client_mixin.